# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwipeJobs API - Tinder/TikTok for Jobs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: Get job feed for swiping (auth optional — guests get all active jobs, authenticated users get personalized feed)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          description: Full-text search across job title, description, and tags
          schema:
            type: string
        - name: jobType
          in: query
          schema:
            type: string
            enum: [fulltime, contract, parttime, internship]
        - name: experienceLevel
          in: query
          schema:
            type: string
            enum: [junior, mid, senior, lead, executive, any]
        - name: remoteType
          in: query
          schema:
            type: string
            enum: [remote, hybrid, onsite]
      responses:
        "200":
          description: Job feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
...
  /employer/me:
    get:
      operationId: getEmployerProfile
      tags: [employer]
      summary: Get employer profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Employer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerProfile"
        "404":
          description: No employer profile found
    post:
      operationId: createEmployerProfile
      tags: [employer]
      summary: Create employer profile (employer onboarding)
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployerProfileRequest"
      responses:
        "201":
          description: Created employer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerProfile"
    patch:
      operationId: updateEmployerProfile
      tags: [employer]
      summary: Update employer profile
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployerProfileRequest"
      responses:
        "200":
          description: Updated employer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerProfile"
  /employer/jobs:
    get:
      operationId: getEmployerJobs
      tags: [employer]
      summary: Get all jobs posted by this employer
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Employer job listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJobsResponse"
    post:
      operationId: createEmployerJob
      tags: [employer]
      summary: Create a new job posting (requires Stripe payment)
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created (pending payment)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJobCreatedResponse"
  /employer/jobs/{jobId}/analytics:
    get:
      operationId: getEmployerJobAnalytics
      tags: [employer]
      summary: Get analytics for a specific job posting
      security:
        - clerkAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAnalytics"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - id
        - name
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          $ref: "#/components/schemas/Company"
        location:
          type: string
          nullable: true
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
        source:
          type: string
        matchScore:
          type: integer
          nullable: true
          description: Match percentage 0-100 based on user preferences
        isSaved:
          type: boolean
        isPaidListing:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string
        expiresAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - company
        - remoteType
        - jobType
        - experienceLevel
        - tags
        - shortDescription
        - fullDescription
        - applyUrl
        - source
        - isSaved
        - isPaidListing
        - viewCount
        - createdAt
...
    JobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - jobs
        - total
        - hasMore
...
    EmployerProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        companyName:
          type: string
        companyWebsite:
          type: string
          nullable: true
        companyLogoUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isVerified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - companyName
        - isVerified
        - createdAt
    CreateEmployerProfileRequest:
      type: object
      properties:
        companyName:
          type: string
        companyWebsite:
          type: string
        companyLogoUrl:
          type: string
        description:
          type: string
      required:
        - companyName
    UpdateEmployerProfileRequest:
      type: object
      properties:
        companyName:
          type: string
        companyWebsite:
          type: string
        companyLogoUrl:
          type: string
        description:
          type: string
    EmployerJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/EmployerJobListing"
        total:
          type: integer
      required:
        - jobs
        - total
    EmployerJobListing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
          nullable: true
        remoteType:
          type: string
        jobType:
          type: string
        experienceLevel:
          type: string
        isActive:
          type: boolean
        isPaidListing:
          type: boolean
        viewCount:
          type: integer
        saveCount:
          type: integer
        clickCount:
          type: integer
        createdAt:
          type: string
        expiresAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - remoteType
        - jobType
        - experienceLevel
        - isActive
        - isPaidListing
        - viewCount
        - saveCount
        - clickCount
        - createdAt
    EmployerJobCreatedResponse:
      type: object
      properties:
        jobId:
          type: integer
        checkoutUrl:
          type: string
          nullable: true
        message:
          type: string
      required:
        - jobId
        - message
    JobAnalytics:
      type: object
      properties:
        jobId:
          type: integer
        title:
          type: string
        views:
          type: integer
        saves:
          type: integer
        clicks:
          type: integer
        clickThroughRate:
          type: number
      required:
        - jobId
        - title
        - views
        - saves
        - clicks
        - clickThroughRate
...
    CreateJobRequest:
      type: object
      properties:
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        location:
          type: string
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
      required:
        - title
        - remoteType
        - jobType
        - experienceLevel
        - shortDescription
        - fullDescription
        - applyUrl
...